Fixed Android build problems.
git-svn-id: svn://0x263a.de/tekuum_src@371 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ sys/android/android_snd.cpp
@@ -123,7 +123,7 @@ public:
 	
 	
 	// Linux driver doesn't support memory map API
-	bool				Lock( void** pDSLockedBuffer, ulong* dwDSLockedBufferSize )
+	bool				Lock( void** pDSLockedBuffer, uint32* dwDSLockedBufferSize )
 	{
 		return false;
 	}
@@ -131,7 +131,7 @@ public:
 	{
 		return false;
 	}
-	bool				GetCurrentPosition( ulong* pdwCurrentWriteCursor )
+	bool				GetCurrentPosition( uint32* pdwCurrentWriteCursor )
 	{
 		return false;
 	}
